requirements.txt
nicegui
uvicorn
gunicorn

requests

azure-core
azure-identity